Add sessions, users and accounts
Following the model used on Campfire, this adds the ability to create
users, and log in. It includes:

- User and Account models. The account is used as a singleton.
- A first-run step which creates the account and the first user.
- Authentication, with DB-backed sessions.
- Avatars, with fallback to SVG initials.

The UI for this is currently just some bare-bones forms to allow signing
up or logging in. We aren't using the user information anywhere in the
UI yet, and we haven't wired up a way to log out.