Include edit forms in arrangement when dragging
If there is a new item showing its edit for when we start a drag, we
need it to be included in the arrangement so we can drag the next item
around it. If we don't do this, there's a "hole" in the list during the
drag operation, which looks broken.